- Enemy now has a damage variable - Player now takes damage according to the damage of the enemy

+++ Shapeshifter/Assets/Scripts/CharacterStats.cs
@@ -17,7 +17,7 @@ public class CharacterStats : MonoBehaviour
         currentHealth = maxHealth;
     }
 
-    public void TakeDamage(int strength)
+    public void TakeDamage(float strength)
     {
         strength -= defence.getValue();
         strength = Mathf.Clamp(strength, 0, int.MaxValue);
